Create promotions_controller.rb
code sample for promotions controller for nail salon app. The nail salon app intends to make the process of making appointments and checking in at a nail salon easier (this is defined by usability) as well as for managers to manage the customers that attend the salon. These controller functions are intended for managers to create promotions for customers.